+++ db/migrate/20170330121215_add_autocalc_default_to_users.rb
@@ -1,0 +1,5 @@
+class AddAutocalcDefaultToUsers < ActiveRecord::Migration
+  def change
+    add_column :users, :autocalc_default, :boolean, default: true, null: false
+  end
+end