+++ db/migrate/20170330125851_add_character_sheet_url_to_character.rb
@@ -1,0 +1,5 @@
+class AddCharacterSheetUrlToCharacter < ActiveRecord::Migration
+  def change
+    add_column :characters, :character_sheet_url, :string
+  end
+end