+++ spec/factories/users.rb
@@ -3,5 +3,7 @@ FactoryGirl.define do
     email { Faker::Internet.email }
     password { 'password' }
     password_confirmation { 'password' }
+
+    autocalc_default { true }
   end
 end