+++ spec/features/trade_logs/manage_trade_log_entry_spec.rb
@@ -6,8 +6,8 @@ RSpec.feature 'Trade Log Entry page', type: :feature, js: true do
     login_as(@user, scope: :user)
     @character = FactoryGirl.create(:character, user: @user)
 
-    @log_entry = FactoryGirl.create(:character_log_entry)
-    @magic_item = FactoryGirl.create(:magic_item, name: 'Staff of Power', rarity: 'rare', log_entry: @log_entry)
+    @log_entry = FactoryGirl.create(:character_log_entry, characters: [@character])
+    @magic_item = FactoryGirl.create(:magic_item, name: 'Staff of Power', rarity: 'rare', log_entry: @log_entry, character: @character)
   end
 
   scenario 'Adding trade log hides old magic item and shows new magic item' do
@@ -40,10 +40,10 @@ RSpec.feature 'Trade Log Entry page', type: :feature, js: true do
 
   context 'Character has trade log entry with magic item' do
     before(:each) do
-      @magic_item = FactoryGirl.create(:magic_item, name: 'Rod of Wonder', rarity: 'legendary', log_entry: @log_entry)
+      @magic_item = FactoryGirl.create(:magic_item, name: 'Rod of Wonder', rarity: 'legendary', log_entry: @log_entry, character: @character)
 
       @trade_log_entry = FactoryGirl.create(:trade_log_entry, traded_magic_item: @magic_item)
-      @received_magic_item = FactoryGirl.create(:magic_item, name: 'Sword of Awesome', log_entry: @trade_log_entry)
+      @received_magic_item = FactoryGirl.create(:magic_item, name: 'Sword of Awesome', log_entry: @trade_log_entry, character: @character)
 
       @character.log_entries = [@log_entry, @trade_log_entry]
     end